Factoring out types and hooks to own directories.
