+++ headless-dropdown-component/src/Components/Dropdown/hooks/useDropdown.ts
@@ -1,0 +1,6 @@
+import { useState } from "react";
+import { DropdownItem } from "../DropdownMenu";
+
+const useDropdown = (items: DropdownItem[]) => {
+  const [isOpen, setIsOpen] = useState(false);
+};
