+++ headless-dropdown-component/src/Components/Dropdown/types/DropdownTypes.ts
@@ -1,0 +1,11 @@
+interface DropdownItem {
+  icon?: string;
+  text: string;
+  description: string;
+}
+
+type DropdownProps = {
+  dropdownItems: DropdownItem[];
+};
+
+export type { DropdownItem, DropdownProps };
